+++ src/hooks/actions/use-msg-actions.ts
@@ -99,8 +99,8 @@ export const useMsgActions = ({
 		folderId,
 		isMessageRead: message.read
 	});
-	const flagDescriptor = useMsgSetFlagDescriptor([message.id], message.flagged);
-	const unflagDescriptor = useMsgSetUnflagDescriptor([message.id], message.flagged);
+	const flagDescriptor = useMsgSetFlagDescriptor([message.id], !!message.flagged);
+	const unflagDescriptor = useMsgSetUnflagDescriptor([message.id], !!message.flagged);
 	const markAsSpamDescriptor = useMsgSetSpamDescriptor({
 		ids: [message.id],
 		shouldReplaceHistory,