+++ src/views/app/detail-panel/edit/tests/edit-view-controller.test.tsx
@@ -9,6 +9,7 @@ import React from 'react';
 import { faker } from '@faker-js/faker';
 import { act } from '@testing-library/react';
 import { Board } from '@zextras/carbonio-shell-ui';
+import { ErrorSoapBodyResponse } from '@zextras/carbonio-ui-soap-lib';
 import { HttpResponse } from 'msw';
 
 import { updateMessages } from '../../../../../store/emails/store';
@@ -22,7 +23,7 @@ import {
 	createAPIInterceptor,
 	createSoapAPIInterceptor
 } from '@test-utils/network/msw/create-api-interceptor';
-import { ASSERTIONS } from '__test__/constants';
+import { buildSoapErrorResponseBody } from '@test-utils/utils/soap';
 import { setupEditorStore } from '__test__/generators/editor-store';
 import { populateMessagesInEmailStore } from '__test__/generators/generateMessage';
 import { EditViewActions } from 'constants/index';
@@ -95,7 +96,7 @@ describe('EditViewController', () => {
 
 		await act(async () => setupTest(<EditViewController />));
 
-		expect(apiCallFlag).not.toHaveBeenCalledWith();
+		expect(apiCallFlag).not.toHaveBeenCalled();
 	});
 
 	it.each`
@@ -106,90 +107,12 @@ describe('EditViewController', () => {
 		${EditViewActions.FORWARD_AS_ATTACHMENT}
 		${EditViewActions.EDIT_AS_NEW}
 		${EditViewActions.EDIT_AS_DRAFT}
-	`(
-		`should not call the getMsg API if the action is $action but the required  message is fully loaded`,
-		async ({ action }) => {
-			getUserSettings.mockReturnValue({
-				attrs: {},
-				props: [],
-				prefs: {
-					zimbraPrefComposeFormat: 'html'
-				}
-			});
-			const messages = populateMessagesInEmailStore({
-				messageGeneratorParams: [{ truncated: false, isComplete: true }]
-			});
-
-			const boardMock = createBoardMock({
-				originAction: action,
-				originActionTargetId: messages[0].id
-			});
-			useBoard.mockReturnValue(boardMock);
-			const apiCallFlag = vi.fn();
-			createSoapAPIInterceptor('GetMsg', messageMock).finally(apiCallFlag);
-
-			await act(async () => setupTest(<EditViewController />));
-
-			expect(apiCallFlag).not.toHaveBeenCalledWith();
-		}
-	);
-
-	it.each`
-		action                                   | isComplete           | isTruncated
-		${EditViewActions.REPLY}                 | ${ASSERTIONS.IS_NOT} | ${ASSERTIONS.IS}
-		${EditViewActions.REPLY}                 | ${ASSERTIONS.IS_NOT} | ${ASSERTIONS.IS_NOT}
-		${EditViewActions.REPLY}                 | ${ASSERTIONS.IS}     | ${ASSERTIONS.IS}
-		${EditViewActions.REPLY_ALL}             | ${ASSERTIONS.IS_NOT} | ${ASSERTIONS.IS}
-		${EditViewActions.REPLY_ALL}             | ${ASSERTIONS.IS_NOT} | ${ASSERTIONS.IS_NOT}
-		${EditViewActions.REPLY_ALL}             | ${ASSERTIONS.IS}     | ${ASSERTIONS.IS}
-		${EditViewActions.FORWARD}               | ${ASSERTIONS.IS_NOT} | ${ASSERTIONS.IS}
-		${EditViewActions.FORWARD}               | ${ASSERTIONS.IS_NOT} | ${ASSERTIONS.IS_NOT}
-		${EditViewActions.FORWARD}               | ${ASSERTIONS.IS}     | ${ASSERTIONS.IS}
-		${EditViewActions.FORWARD_AS_ATTACHMENT} | ${ASSERTIONS.IS_NOT} | ${ASSERTIONS.IS}
-		${EditViewActions.FORWARD_AS_ATTACHMENT} | ${ASSERTIONS.IS_NOT} | ${ASSERTIONS.IS_NOT}
-		${EditViewActions.FORWARD_AS_ATTACHMENT} | ${ASSERTIONS.IS}     | ${ASSERTIONS.IS}
-		${EditViewActions.EDIT_AS_NEW}           | ${ASSERTIONS.IS_NOT} | ${ASSERTIONS.IS}
-		${EditViewActions.EDIT_AS_NEW}           | ${ASSERTIONS.IS_NOT} | ${ASSERTIONS.IS_NOT}
-		${EditViewActions.EDIT_AS_NEW}           | ${ASSERTIONS.IS}     | ${ASSERTIONS.IS}
-		${EditViewActions.EDIT_AS_DRAFT}         | ${ASSERTIONS.IS_NOT} | ${ASSERTIONS.IS}
-		${EditViewActions.EDIT_AS_DRAFT}         | ${ASSERTIONS.IS_NOT} | ${ASSERTIONS.IS_NOT}
-		${EditViewActions.EDIT_AS_DRAFT}         | ${ASSERTIONS.IS}     | ${ASSERTIONS.IS}
-	`(
-		`should call the getMsg API if the action is $action, the required message $isComplete.desc complete and $isTruncated.desc truncated`,
-		async ({ action, isComplete, isTruncated }) => {
-			const messages = populateMessagesInEmailStore({
-				messageGeneratorParams: [{ truncated: isTruncated.value, isComplete: isComplete.value }]
-			});
-
-			const boardMock = createBoardMock({
-				originAction: action,
-				originActionTargetId: messages[0].id
-			});
-			useBoard.mockReturnValue(boardMock);
-			const soapMessage = getSoapMailMessage(messages[0].id);
-			const getMsgInterceptor = createSoapAPIInterceptor<GetMsgRequest, GetMsgResponse>('GetMsg', {
-				m: [soapMessage]
-			});
-
-			await act(async () => setupTest(<EditViewController />));
-
-			const getMsgRequest = await getMsgInterceptor;
-
-			expect(getMsgRequest).toEqual(
-				expect.objectContaining({
-					m: expect.objectContaining({
-						id: messages[0].id
-					})
-				})
-			);
-		}
-	);
-	it('should call getMsg API when the preferred body format is not available', async () => {
+	`(`should call the getMsg API if the action performed is $action`, async ({ action }) => {
 		getUserSettings.mockReturnValue({
 			attrs: {},
 			props: [],
 			prefs: {
-				zimbraPrefComposeFormat: 'plain'
+				zimbraPrefComposeFormat: 'html'
 			}
 		});
 		const messages = populateMessagesInEmailStore({
@@ -197,26 +120,16 @@ describe('EditViewController', () => {
 		});
 
 		const boardMock = createBoardMock({
-			originAction: EditViewActions.REPLY,
+			originAction: action,
 			originActionTargetId: messages[0].id
 		});
 		useBoard.mockReturnValue(boardMock);
-		const soapMessage = getSoapMailMessage(messages[0].id);
-		const getMsgInterceptor = createSoapAPIInterceptor<GetMsgRequest, GetMsgResponse>('GetMsg', {
-			m: [soapMessage]
-		});
+		const apiCallFlag = vi.fn();
+		createSoapAPIInterceptor('GetMsg', messageMock).finally(apiCallFlag);
 
 		await act(async () => setupTest(<EditViewController />));
 
-		const getMsgRequest = await getMsgInterceptor;
-
-		expect(getMsgRequest).toEqual(
-			expect.objectContaining({
-				m: expect.objectContaining({
-					id: messages[0].id
-				})
-			})
-		);
+		expect(apiCallFlag).toHaveBeenCalled();
 	});
 
 	it("shouldn't unmount the editor when the message is updated", async () => {
@@ -236,7 +149,10 @@ describe('EditViewController', () => {
 				}
 			]
 		})[0];
-
+		const soapMessage = getSoapMailMessage(message.id);
+		const getMsgInterceptor = createSoapAPIInterceptor<GetMsgRequest, GetMsgResponse>('GetMsg', {
+			m: [soapMessage]
+		});
 		const boardMock = createBoardMock({
 			originAction: EditViewActions.REPLY,
 			originActionTargetId: message.id
@@ -244,7 +160,7 @@ describe('EditViewController', () => {
 		useBoard.mockReturnValue(boardMock);
 
 		await act(async () => setupTest(<EditViewController />));
-
+		await getMsgInterceptor;
 		expect(screen.getByRole('button', { name: /send/i })).toBeVisible();
 
 		expect(updateBoardContext).toHaveBeenCalledTimes(1);
@@ -269,5 +185,42 @@ describe('EditViewController', () => {
 		});
 
 		expect(screen.getByRole('button', { name: /send/i })).toBeVisible();
+	});
+	it('should render a loader when the message is not available', async () => {
+		const messages = populateMessagesInEmailStore({
+			messageGeneratorParams: [{ isComplete: false }]
+		});
+
+		const boardMock = createBoardMock({
+			originAction: EditViewActions.REPLY,
+			originActionTargetId: messages[0].id
+		});
+		useBoard.mockReturnValue(boardMock);
+		const errorResponse = buildSoapErrorResponseBody();
+		const getMsgInterceptor = createSoapAPIInterceptor<GetMsgRequest, ErrorSoapBodyResponse>(
+			'GetMsg',
+			errorResponse
+		);
+
+		await act(async () => setupTest(<EditViewController />));
+		await getMsgInterceptor;
+		expect(screen.getByTestId('EditViewControllerLoader')).toBeVisible();
+	});
+
+	it('should not render a loader when the message is available', async () => {
+		const messages = populateMessagesInEmailStore();
+		const soapMessage = getSoapMailMessage(messages[0].id);
+		const getMsgInterceptor = createSoapAPIInterceptor<GetMsgRequest, GetMsgResponse>('GetMsg', {
+			m: [soapMessage]
+		});
+		const boardMock = createBoardMock({
+			originAction: EditViewActions.REPLY,
+			originActionTargetId: messages[0].id
+		});
+		useBoard.mockReturnValue(boardMock);
+		await act(async () => setupTest(<EditViewController />));
+		await getMsgInterceptor;
+
+		expect(screen.getByTestId('edit-view-editor')).toBeVisible();
 	});
 });